requirements.txt
Django==4.2.5
psycopg2-binary>=2.8
requests==2.31.0
beautifulsoup4==4.12.3
numpy==1.26.0
scipy==1.12.0
gensim==4.3.2
pandas==2.2.0
transformers>=4.0.0
tensorflow>=2.0.0
torch==2.2.0
newspaper3k==0.2.8